Turn off step nav for rules in `pages_related_to_step_nav`
Reverse engineer a step by step from publishing api and save it
as a StepByStepPage in collections-publisher.

Step by step navigation is shown by default on content pages, unless it is specifically turned off. Those pages are
listed as `pages_related_to_step_nav` in the step by step.

This change checks whether any of the internal links scraped from the step are in the `pages_related_to_step_nav` list
and set show navigation to false if they are.

Reverse engineering a step by step is the first step in allowing publishers to discard drafts of previously published
step by steps.
Rather than creating Editions in collections-publisher, we are simply going to replace the latest version
of the step nav with the last published version from publishing-api.