Rescue error from missing batch report
Link-checker-api throw a 5xx error if `get_batch` is called for a batch report that doesn’t exist.

This isn’t usually a problem in Production, but step by step pages, and their associated link reports are copied
to integration and staging each night. This means that if when you try to view the step by step in those environments, it’s
querying link-checker-api with a batch id that was created in (and probably only exists in) production. This causes link-checker-api to
fail with a 5xx.

We should update link-checker-api to fail more gracefully if it can’t find a requested batch, but until then, allow collections-publisher
to load step by steps even if the link report details cannot be found.

Error in [sentry](https://sentry.io/govuk/app-collections-publisher/issues/686572073/?query=is:unresolved)