Update edition number of change notes on publish
After a step by step is published, get the latest user_facing_version from publishing-api and store it against the change notes.

This will allow us in the future, to use the user_facing_version stored in the change notes and it marry to the edition stored in publishing-api to allow publishers in the future to see what was changed and why.

For now though, if a change note has a edition number, we will assume it belongs to a previous edition. However, if a change note does not have an edition number, we’ll assume it belongs to the latest draft, and delete it if the draft is ever discarded.