Set draft_updated_at time to published_at from step by step
Reverse engineer a step by step from publishing api and save it
as a StepByStepPage in collections-publisher.

We determine whether a draft is in progress if the draft_updated_at time is different to the published_at time.
Delays in publishing mean that the public_updated_at time from publishing-api doesn't always match the published_at
time in the step by step. To get around this discrepancy, the draft_updated_at time is being reset to match
published_at time, as this is time it would have originally been set to before the draft that is being discarded was
created.

Reverse engineering a step by step is the first step in allowing publishers to discard drafts of previously published
step by steps.
Rather than creating Editions in collections-publisher, we are simply going to replace the latest version
of the step nav with the last published version from publishing-api.