Fix the delete step text
The `User deletes a step` feature test was failing with the following error:

```
ActiveRecord::RecordInvalid:
       Validation failed: Title can't be blank, Base path can't be blank, Content can't be blank, Publishing app can't be blank, Schema name can't be blank
```

The factory being used created two steps, ` step` and an `or_step`. The `or_step` did not have any content, so when the first `step` was deleted and the navigation_rules were updated, they threw an error.

I have no idea why the test didn’t fail when it was first added.