Store the latest edition number in change notes
Sometimes content designers need record change notes for step by steps that have already been published.

Getting the edition number and storing the edition number for these change notes allows us to keep the version
history of the step by step up to date.

The default behaviour of not storing the edition number of the step by step in a change note, for step by steps in a draft
state is not being changed.