Get version number of latest published edition
Functionality is being added to allow publishers to discard drafts of previously published content.
When these drafts are discarded, the data in the step by step will be replaced with the content of
latest published version held in publishing api.

By default, `get_content` retrieves the latest version regardless of publication state. Though we will
be sending a discard-draft request to publishing-api, this request may not have completed by the
time we are ready to revert the step-by-step, so we should make sure that we are over-writing it with
the correct version of the content.

However, we can specify which version of content we want by passing in the `latest_edition_number` as
the `version` when we are ready to replace the data:

e.g. `Services.publishing_api.get_content(content_id, version: 6)`