Discards drafts cleanly before destroying model
This commit ensures that when a Step By Step is deleted, the change notes without an edition are deleted first. This ensures database integrity is maintained. It fixes a foreign key error that was being thrown when a draft guide that had changenotes was discarded. This is because the guide was discarded before the changenotes, which orphaned the changenotes and caused the error. The fix is implemented by calling a before_destroy action that deletes the changenotes associated to that draft.