Add a new route to revert published changes
This is the route that will be called from the “Publish or Delete” page when the user discards a draft.

It’s called “revert”, because that’s what’s happening. The draft won’t be discarded in the traditional sense
i.e. we’re not dropping a row from the database. Rather we’re overwriting (“reverting”) the data in the step
by step.