Adds method to assign current user to current draft
This commit adds a small method that assigns whoever the current user is to the current draft when it is updated. This happens if a new guide is created or an existing guide is edited in any way. It should ensure that at the very least users can see who is currently responsible for a draft. The meat of this commit is updating areas where this method is called and updating the tests as well.

I'm going to extend this in a future PR so that a change note is generated. This will create a "Chain of Custody" that describes who's been assigned the draft.