Adds function to generate change note when user creates new draft
This commit adds tests and a method to generate an InternalChangeNote with fixed phrasing when a user creates a new draft. It does not yet allow for users to assign drafts to each other. This change is being made so that users can see who began the most recent draft and then follow up with them.