Changes #update_assigned_to method
This commit alters the StepByStepDraftUpdateWorker#update_assigned_to method. Previously, it only saved the assignee the first time the draft was created. Now it is called whenever a user saves an in-flight draft. It will only be called once per user, as it generates a changenote each time. This last limit is to ensure the changenote history isn't flooded with the same message multiple times.