Cope with trailing whitespace in step markdown
At present, steps with multiple trailing lines will create an empty paragraph once
parsed.

This breaks Google's rich snippet parser for HowTo schemas, so we should fix
it.

Note that it doesn't have any visible effects on the page as we seem to
defend against that in the step by step component.

https://collections-publisher.publishing.service.gov.uk/step-by-step-pages/66/internal-change-notes