Check is secondary content is a smartanswer
Smartanswers are comprised of two content items, the landing page, and
the question and answer pages.

Only smartanswer landing pages are linked to from a step by step by
content designers, but to ensure that users can still see the step by
step when they are going  through the smartanswer questions, the second
half of the smartanswer is automatically added. This behaviour needs to
be replicated for secondary content links.

The first step is duplicate the `smartanswer?` method in NavigationRule.