Add done pages to secondary_to_step_nav
Similar to part_of_step_nav, this adds the smartanswer done page
content_ids and service done page content_ids  to the
secondary_of_step_nav links in the payload for publishing-api if the
smartanswer start page content_id or service start page content id is
already in that list.

The done page links are being found and added to the publishing-api
payload, right at the last minute, and are not being stored as secondary
content links to remove the overhead of having to keep them up to date
if the start page is removed from secondary content links.

As both NavigationRule and SecondaryContentLink respond to
`smartanswer?` some of the existing code for finding done pages can be
reused.