Revert secondary content when draft is discarded
Secondary content links are pushed to publishing_api as "edition" links, and not as document
links, so they are tied to the version of the step by step page.
In collections publisher we discard drafts of step by steps by getting the previous version from
publishing-api and using that to repopulate the step by step locally. As the links are tied to the
version, we can revert them too.

The feature test was updated to add a `links` hash to fix error:

```
Failure/Error: payload_from_publishing_api["links"][:pages_secondary_to_step_nav] || []

     NoMethodError:
       undefined method `[]' for nil:NilClass
```