Make `published_step_by_step_page` return StepByStepPage in live state
Previous version had different values for `published_at` and `draft_updated_at`,
meaning the StepByStepPage.state returned `unpublished_status` rather than
`live_status`. The name of this factory was therefore quite misleading.

In the next commit, I use this factory to test a feature that is dependent
upon the StepByStepPage being in a live state, and rather than add a new
factory with an overly long name to disambiguate the two, I've refactored
the original factory to fix the underlying issue.

All of the tests continued to pass apart from a test for the 'Discard changes'
functionality. In this test I have now made it explicitly clear that it is
testing a published step by step page that contains unpublished changes.