Remove "future" check for scheduling status
In the next commit `scheduled_for_publishing?` will be used to determine whether
a scheduled job is still valid and should still be sent for publishing.

To do this the future check needs to be removed. This to avoid conflicts with
datestamps where daylight savings, and differences in milliseconds might cause a
job not to be published.

For example, the scheduled_at time is set to 12:00:00, but time now is 12:00:05.
`future?` will return false, even though for all intents and purposes, the times are
the same.

Rather than relying on `future?`, the `scheduled_at` time will be cleared in a later
commit when the step by step is published.