Only publish a step by step if it's been scheduled
Sidekiq job can't be removed from the queue once added, so we need to
make sure that only valid requests are sent to publishing-api.

Uses the `scheduled_for_publishing?` to make sure the request is still
valid. This check expects the step by step to have a draft, so a new
`draft_step_by_step_page` has been added to the tests.