Unassign the user when step by step is published
Change notes are only generated when a draft is saved by a "new user".
[This][1] was added to prevent a change note from being added every time
the user clicks the "save" button.

Unfortunately, this meant that a change note wasn't being added if the
same user who had previously published the step by step made new
changes. To solve this problem the user is being unassigned when the
step by step is published.

The feature tests alias one of the existing methods, to make the Given,
When, Then flow more readable

[1]: https://github.com/alphagov/collections-publisher/blob/master/app/workers/step_by_step_draft_update_worker.rb#L27