Show error message when secondary link slug doesn't exist
Fixes a regression whereby no error message is displayed if a
slug for a step by step's secondary link doesn't exist.
The error wasn't showing as it's outputting to flash[:danger],
which was no longer getting displayed in admin_layout.html.
I now output it to flash[:alert] in line with the recent design
system changes.

The regression wasn't caught in tests as the test didn't assert
that the page contained content (it was missing an 'expect()').