Move the logic of showing elements from the view to the model
The logic of showing elements on the Delete or publish page is going to be more
complicated when we ‘lock’ the step by step page for editing. Moving the logic
into the model makes it possible to test it using unit tests. This also means
that the logic in the view will be simpler.