Adds error messages at the top of the page
In addition to inline error messages, we want error messages to appear at
the top of the page, which is the pattern that Content Publisher uses,
meaning you get both immediate and specific feedback.

DatetimeParser issues are split into `:schedule_date` and `:schedule_time`
namespaces so they can be easily mapped to the inputs that caused the errors.
However Rails automatically prefixes error messages with their namespace,
so errors like "Enter a valid date" become "Schedule date Enter a valid date".
So I've added to the `:base` namespace to avoid any prefix.