Ensure change notes can be added to scheduled step by steps
We lock down a lot of behaviour when a step by step is scheduled,
but it should still be possible to add internal change notes.

Have refactored the "User leaves a change note" scenario to be a
more readable test as part of this.

Trello: https://trello.com/c/i5J47svc/89-lock-a-step-by-step-that-has-been-scheduled-for-publication-on-the-change-notes-page