Test can still check for broken links
When a step by step has been scheduled for publishing, publishers should
not be able to make changes to the step by step, so that draft changes
can't be published unexpectedly.

Publishers should still be able to check for broken links as this
doesn't change the content of the step by step. The expectation is that
if publishers want to fix the links, they need to first unschedule
publishing, fix the links, then reschedule publishing.