Replace accordion with a timeline entry
Replace the accordion of each history entry with a timeline entry element, consistent with the tested pattern in Content Publisher and replace 'change note' with 'internal note' to avoid confusion with public change notes.

Make the form visible all the time. This means that the JavaScript and SCSS used to reveal the form (bindCancelAddChangeNoteLink) is not needed anymore.