Set the time zone to London
Config taken from: https://github.com/alphagov/content-publisher/commit/4ae1c4ed683bf0b24eb11d135aebd3cdd1a8ffa6

We store dates in the database in UTC, which means that when we're displaying
dates and times in the views they're an hour out of date during daylight saving
time. This sets the app's time zone to 'London' so that ActiveRecord
automatically converts dates and times to the UK's current time zone, whether
it's GMT or BST.

Also updates a test that was trying to get around this problem by converting the dates checked to "UTC +00:00" and
another test that was setting the date to "London" before doing the comparison.