Hide schedule button if step by step can't be published
If the step by step is not ready to be published, because it doesn't have any steps,
or some of it steps don't have content, then we shouldn't allow scheduled publications either.

This reorders the items on the publish-or-delete page to better reuse the existing `can_be_published?` check.
This doesn't affect the order of the buttons on the page.