Use the metadata component
Creates and app-side object (we structure the SCSS files following the ITCSS arhictecture) to be used for actions and metadata. By doing so we don't need the `step-by-step-actions__list` anymore and we  use it consistently for any side container.

Because admin metadata is styled differently than the public facing metadata we need a set of overwrites to increase the font size and align the elements as per design

We then use the `app-side` in 'show' view to wrap actions and metadata component