Bypass changelog when publishing step by steps for the first time
It does not make sense for a content designer to choose a Major or
Minor update and accompanying note when first publishing a step by
step.

For now, we can default to treating it as a minor update without a
changenote (if we had chosen major update, we would have to provide
a changenote). We will worry about rendering a different changenote
message in the next commit.

I've had to change step_by_step_pages_controller_spec.rb to test
major/minor updates against already published step by steps, as it
is no longer an option to provide a changenote for a first-time
publish.

I've also had to move the `|| ""` logic in publish_intent.rb, as
otherwise calls to @publish_intent.change_note.empty? fail (NilClass)
when we're not passing a change_note.