Filter results using StepByStepFilter class
* Changes the ids of the text and select input so that they no longer
trigger the javascript filtering. The javascript will be removed in the
next commit.

* Adds a form and "Filter" button to make a request to
`StepByStepFilter`.
`form_tag` is being used rather than `form_for` because the form is not
associated with a model object. Unfortunately, `form_tag` automatically
adds a `utf8 = ✓` param to the querystring force Internet Explorer (5,
6, 7 and 8) to encode its parameters as unicode.
It's very unlikely that any of our users will be using old versions
of IE, so rather than add `utf8` to the permitted params, it's being
explicitly omitted in the form using the `enforce_utf8` flag.

For more information, see:
https://stackoverflow.com/questions/4104474/rails-3-utf-8-query-string-showing-up-in-url/28201543#28201543
https://stackoverflow.com/questions/3222013/what-is-the-snowman-param-in-ruby-on-rails-3-forms-for/3348524#3348524

* Adds a "Clear all filters" link that reloads the page thereby
resetting the selected filters.

* The status select values are still being hardcoded. The text and value
of the status is also being defined in the `StepByStepPage` model.
The config here and in the model should be moved to a common
helper in a future enhancement when we start explicitly storing the
status, as then we'll know how best to consolidate the status text.