Display title/slug/intro/description with summary_list
Adds new test scenario so we can begin to assert the presence
of the different sections of the new step by step page. I've
added tests for the metadata section - this is already somewhat
covered in other tests but those are specifically about checking
that the 'Last updated' field shows the author name, so think
there is still value in adding here.

Deliberately not testing the 'edit steps' portion of the page
as its content is going to change in a future PR.