Set status to draft when step-by-step unpublished
We're setting the status to draft so that unpublished step-by-steps will
appear in the filtered results on the index page.

Before, when the status was being calculated, unpublished step-by-steps
didn't return a value for status. Now that a status is required, the
status is being set to "draft" by default, but we should consider adding
an "unpublished" status(es) in the future.