Add rake task to set status of step-by-steps
"update_column" bypasses the model validations so it is being used in
the tests to simulate the state of existing step-by-steps that won't
have anything set for the status.

It's not possible to use the existing scheduled_for_publishing?` method
from the model because it has already been updated to include a check
for `status == "scheduled"`.