Fix redirect URL when step is edited
Clicking 'Edit' takes you to 'steps/step_id/edit', but after editing you are
taken to 'steps/step_id' (no '/edit'), even though it appears to the user to
be the same page. I believe this was an oversight and the intention is that
the user should remain on the same URL.

Have changed the steps so that they are more indicative of the user journey now
that the UI has changed.
I've also made the 'then_the_step_is_deleted' step less specific so that if the
markup changes, the test won't falsely pass.