Improve 'editing title/slug/intro/description' test
Old test implied a step by step must have a step before you can
edit its information, but that's not the case. It also only
validated the update to the Title field (probably because updating
redirects you to the overview page, which only displayed the Title
until the recent UI change), so now we check that all of the fields
have updated.