Add "review_requester_id" column
This is the uid of the person who submitter for 2i. The uid is being
stored rather than the name to:

1. Make sure we always display the correct user name. The name of the
user can change, but the uid won't.
2. Make is easier to make comparisons with other users. Two users could
have the same display name, but they won't have the same uid.
3. The uid is indexed in the User table, so these seems to be the best field
to use to define the foreign key relationship.

Unfortunately, rails doesn't like numbers in column names, so "2i"
couldn't form part of the column name.