Add additional change note step to scheduling workflow
Step-by-steps that are published immediately can have a change note that is emailed
to subscribers (by marking the change as major).  Scheduled step-by-steps could not.

This adds an additional step to the scheduling workflow (prior to selecting the date)
allowing the change type to be selected and a change note to be entered (for major
updates only).

The update type and change note are persisted through the workflow using session
variables.  We do not want to send the change note to Publishing API when it is
initially entered, since the value is pointless until the step-by-step is actually
scheduled for publication.

Validation errors on the date/time will continue to be displayed on the same page
as the date selector.