Present the scheduled update type and change note to Publishing API
If a major change (and therefore a change note) is scheduled, the value is
stored in the database until the point that the change note is actually
scheduled.

This change retrieves these details from the database and appends it to
the Publishing API payload.

A new call to StepNavPublisher.update(@step_by_step_page) has been added
since we intend to remove the concept of PublishIntent from scheduling.