Add reviewer_id column
This follows the pattern set by `review_requester_id` to set it to the `uid` of the user.
This is to:

1. Make sure we always display the correct user name. The name of the
user can change, but the uid won't.
2. Make is easier to make comparisons with other users. Two users could
have the same display name, but they won't have the same uid.

Also, the uid is indexed in the User table, so these seems to be the best field
to use to define the foreign key relationship.

See: https://github.com/alphagov/collections-publisher/commit/0856749acefbe02895f261b953f916fd0f3edc7d#diff-d319dc3e044a472df059bd1336b9705d