Set 'draft_updated_at' for all factory-produced step by steps
The factory produces step by steps that are in 'draft' status, so
they _should_ have a 'draft_updated_at' value anyway.

Many tests rely on the 'can_be_published?' method, which will in
(in a later commit) perform a check to see if links have been
published more recently than the 'draft_updated_at' field, so
tests will fail if this is not set.

I've had to set the fixture in the `can_be_published?` test to a
published one, now that the default draft one has a draft_updated_at
field. Setting it to a published one means draft_updated_at is equal
to published_at, meaning can_be_published? is false.