Show multiple required actions inside the inset prompt
It is a better user experience to see all of the problems rather than
dealing with them one at a time. It is now possible to see errors about
adding content to your steps as well as checking for broken links.
It is not possible, however, to see errors about adding content to your
steps and adding at least one step, as the former is simply a specialised
form of the latter.

I've wrapped each link in a `<li>` inside a `govuk-list` so that the correct
styling is applied.