Combine check for publish/schedule buttons into one step
We're essentially testing the same thing: that the user cannot publish the
step by step. Whether that is immediate or delayed is irrelevant.
This composite step reads a little more clearly and less verbose than it did
before, especially as it is a common thing to test in several scenarios.