Add public change note to internal note on scheduled publish
When a step-by-step is published immediately, the public change note is added
to the internal change note.  This matches that behaviour for scheduled changes.

Note that the change note is now added before the step-by-step is marked as
published (but after it is actually published) since the process of marking
as published clears the public change note from the database.