Move 'steps must have content' message to overview page (inset-text)
This copies over the text and guard clause from publish-or-delete.html,
which was removed in https://github.com/alphagov/collections-publisher/pull/734
This section is likely to get quite long, so I've encapsulated it in its
own partial.

I've also set the background colour of the inset prompt, as this is in
the design but not provided by the component.