DRY-up tests for review controller
Rather than having an individual test for various combinations of permissions,
we now pre-define the required permissions for each action then loop through
and test removing each.

Note that we do not want to test removing the `signin` permission, since removing
this permission causes the users to be redirected to Signon, rather than being
denied access.