Implement 'broken_links_found?' helper method
This will be used in the next commit as part of the 'can_be_published?'
condition. Note that I had to include GDS API TestHelpers into the
LinkChecker module here, as it doesn't exist in
'step_by_step_page_spec.rb'. It must already have existed in
StepNavSteps (where it was first used), so its absence wasn't an
issue there.