Remove margin-top for inset text
The inset text component is used in app with the role of flagging required actions before publishing. It sits at the top of the summary tab content before the summary list and it should be aligned with the actions block on the right when displayed on wide viewports. In order to align them we need to remove the margin-top from the inset text component. We usually only set margins at the bottom of each component - this one being an exception as it's expected to sit in a 'prose' context, surrounded by text.