Users can submit for 2i review
In a draft state a content publisher should be able to submit a step by step for 2i review. Along with this change we're introducing a new test suite: contextual action button tests for step by steps to ensure the correct actions are being listed for each state.

This new test suite has a simple aim: to denote exactly which actions
are available (and whether they are Primary, Secondary or Tertiary),
given a step by step is in a particular state and being viewed by a
particular author.