New test suite: contextual action button tests for step by steps
As we're implementing 2i workflow into step by step publication, and
there is the possibility of 'fact check' workflow thereafter, our test
suites are becoming more difficult to maintain and less all-encompassing
of all possible states and use cases.

This new test suite has a simple aim: to denote exactly which actions
are available (and whether they are Primary, Secondary or Tertiary),
given a step by step is in a particular state and being viewed by a
particular author.