Move StepLinksForRules publishing api methods to PublishingApiHelpers
In the next commit, we'll write an integration test which requires
making an edit to a step in a step by step, which triggers updating
downstream. Currently, this would raise the following error:

Validation failed: Title can't be blank, Base path can't be blank, Content can't be blank, Publishing app can't be blank, Schema name can't be blank
./app/services/step_links_for_rules.rb:114:in `block in add_rules'

So we'll need to stub publishing api, and currently there is
no easy way to do that as the logic we need is coupled to
StepLinksForRules. This commit moves it somewhere central.
I've also created a StepLinkFixtures to avoid moving unrelated
code to PublishingApiHelpers whilst making the fixtures available to
other files.