Update StatusPrerequisiteValidator to allow non-status attributes to be updated without changing a status
It was impossible to schedule 2i approved step by step and the user saw
'approved_2i, requires a draft, a reviewer and for status to be in_review,
scheduled or published' validation error. This occurred because
schedule_datetime method tired to update a single attribute scheduled_at, while
the status remained approved_2i. But the validation required the status to be
in_review before it could be set to approved_2i.