Migrate existing steps to ensure they each have a position
I've added code which fixes the position of new step by step steps,
but we need to deal with the possibility of an existing step by step
that has not yet been reordered. In this case, no steps would have a
'position' and the 'make_step_positions_sequential' would not be able
to sort by 'position'.