Don’t prompt to check for broken links when step by step is approved_2i
Setting `draft_updated_at` to `now` triggered the "Check for broken links" prompt
after the SBS is unpublished. We are showing the prompt when SBS is
a draft and submitted_for_2i so given that the reviewers also check the links
manually, we decided to assume that the links are correct at that point and not
to show the prompt when SBS is approved_2i.