Use translations file to store humanised statuses
This lays the groundwork to have the presentation of a status be
different from the result of calling `.humanize` on a status string.

I decided to use the translations as that seemed an appropriate place
for view concerns. As there is not specific views for this data I put in
a grouping named after the model.

I've removed the StepByStepFilter::Options class as this seems to be
somewhat unnecessary for what is essentially view logic. Initially I
tried just cutting down the tests on this but as the conversations
surrounding testing it seemed to require more and more code it became
apparent to me that this is something rather trivial that is building up
a bunch of infrastructure around it and moving the logic into a view
simplifies those concerns.