Make `all_links_in_content` return a single layer array
Previously it returned an array of 1-element arrays, which meant
having to call `flatten` before select/map, or else quering the
array internals when iterating over it (e.g. `href[0] =~ ...`)

We now call `flatten` on the response before returning it, making
it cleaner to work with. I've taken the opportunity to refactor
the `relative_paths` method to be easier to read.