Do not revert step by step to draft under certain circumstances
Sometimes a reviewer or review requester will want to make minor changes
without getting stuck in a 2i review loop.  This change prevents the status
from changing to 'draft' when a step by step that is currently submitted for
2i or in 2i review is edited.  If a step by step has been approved, then the
status is returned to 'draft' and another 2i review is required.