Keep content designers on same page when editing a step
Content designers like to make small changes, save steps, refresh their preview
and repeat. Therefore there was an additional overhead involved in navigating
back to the step that is being edited, when making multiple small changes to it.

A product decision has been made to keep content designers on the 'Edit Step' page
when updating a step, to enable them to make multiple changes more easily. This
reverses logic added in https://github.com/alphagov/collections-publisher/pull/759

NB, we DO want to redirect to the overview page when CREATING steps, but we want
to stay on the edit page when updating existing steps.

Trello card: https://trello.com/c/qqOsgr1W/117-change-app-behaviour-to-not-redirect-user-to-summary-page-after-saving-a-step