Add contextual guidance page
This adds the contextual guidance page, links to it from SBS's that are in review and tests for same. This will provide a place where 2i reviewers can go to see guidance on what the intended use of attributes are.

Trello - https://trello.com/c/K4Uyg9m6/127-add-a-view-giving-some-contextual-guidance-to-2i-reviewers