Remove "Fix broken links" prompt
In many cases step-by-steps contain links to draft content or other
draft step-by-steps. It's also possible to for draft step-by-steps to be
linked to each other. As Link Checker API doesn't handle draft content,
and can only check links that are publicly accessible, publishers could
end up in a chicken or egg situation where they can't publish any
step-by-steps.

By removing this requirement and the accompanying prompt we are trusting
publishers to know which  broken links in the link report can be safely
ignored.