Update status on page refresh after update
This commit makes sure the status of the step by step is in the correct
state after an edit has been made. Previously, the code to update the
state was inside a sidekiq worker which may not have completed upon the
new page being loaded. This lead to wrong links being displayed until
the worker has processed the job.

I decided to create the UpdateStepByStep service even though being very
simple due to the amount of test replication that would need to be added
without it (to each controller action that uses the added
functionality).

Trello:
https://trello.com/c/L4IbSFuv/111-status-doesnt-update-to-draft-until-page-is-refreshed