Add production database configuration
This was previously missing meaning that in a production environment,
Rails was only picking up the configuration from the `DATABASE_URL`
environment variable. This meant that the configuration in the `default`
block wasn't being used in production.

This wasn't a problem before but in Rails 6 the encoding default was
changed from `utf8` to `utf8mb4` which requires a more recent version of
MySQL to use. This was causing the app to break during the end-to-end
tests when it tries to re-create the database.