Create CoronavirusPage model
- Currently, the coronavirus model will only be associated with
the sections/accordions. But as we move more of the content out of the
the yaml file on govuk-coronavirus-content and into the database, it
will grow.
- It also is only representing the landing page at this stage.
But we can add a type column at a later date so it represents
all the coronavirus hub pages, which will be linked to from the
coronaviruspages controller index page.