Move livestream logic
- When we update the livestream, we create a payload by fetching
the livecontent item, turning it into a hash, and combining it with
the livestream data persisted in the database. When we make content
changes, we fetch the github yaml file, turn it into a hash and
combine it with the coronaviruspage data persisted in the database.
- This commit moves the logic that combines the livestream data into
the payload out of the Coronavirus Page Presenter and into the
content builder class and livestream updater.
- This will make it easier to remove when we need to.
- Also simplifies tests which were testing the first time the livestream
object is initialised - now it's in the database we don't need to check
that sort of thing.