Add temp rake task to redirect legacy signups
Adds a temporary rake task to create redirects for specialist sector
email signup URL's.

The current behaviour is for these specialist sector email-signup URL's
to be sent to collections which then redirects to email-alert-frontend.

This commit creates a redirect for each specialist sector email signup
URL which points to the relevant parent page rendered by collections.

Depends on:
https://github.com/alphagov/collections-publisher/pull/1102

---

Trello:
https://trello.com/c/Fi2Z2zDy/399-redirect-and-remove-legacy-email-signup-routes-for-collections