Fix expected link in scheduling test
When a step by step has been scheduled for publishing, the only thing that can be changed is the linked taxon.
Everything else should be view only, including secondary content and sidebar navigation.

I'm not sure how/if these tests were passing before, but they've been updated to look for the correct link.