Add migration to store featured link
We need to allow users to set one link as a featured link in an
accordion (sub-section).

Links in the accordions are just stored as markdown in a text field, so
there is no way to highlight them as "featured". Instead, the base_path
of the featured link will be stored as a separate attribute.

Validations will be added to make sure that:

1. The featured_link does appear in the content of the accordion.
2. The value being added matches one of the links in the accordion.