Set link to featured in publishing-api payload
Sets a link in an accordion to be a featured link in the payload sent to
publishing-api if it matches the link url set as the featured link.

In the frontend, featured links are expected to have a description.
Rather than ask the user to input this as well, we are instead looking
up and grabbing the existing description for content item from
publishing-api.