Revert featured link when draft is discarded
When the draft is discarded, we need to revert the featured link to the
currently published one. To do this we need to find the url of link in
the accordion that has "featured_link" set to `true` and use it to
re-populate "featured_link" in the SubSection model.

Discarding drafts is kicked of by the ModelBuilder.

The ModelBuilder calls the SectionPresenter with the data from
publishing-api, which in turns passes this data to the
SubSectionProcessor to parse the data.

As the link needs to be parsed from the content, content and
featured link need to be returned at the same time. This means that the
object being returned needed to be changed to from a single array to a
hash containing the original array plus a new featured link attribute.

We don't need to make any changes to the ModelBuilder as it just
overwrites the existing database entry with the output from
SectionPresenter.