Add a service to remove markdown from content
Some the content in the timeline will have markdown in it, this will
most likely be in the form of links. We need away to remove these links
from the content, whilst not losing the link text.

Uses Kramdown to convert the markdown to html. This then makes it easy
to strip html from the content in the usual way.

strip_tags returns a newline at the end of the content it renders, so
squish is being used to remove it.

Kramdown is already being added as a dependency of
govuk_publishing_components. It's being added as an explicit dependency
here in case that changes, but won't need to be managed as an extra
dependabot dependency.